crux_core
Module bridge

Structs

Bridge

Bridge is a core wrapper presenting the same interface as the Core but in a serialized form, using bincode as the serialization format.

BridgeWithSerializer

A bridge with a user supplied serializer

Request

Request for a side-effect passed from the Core to the Shell. The EffectId links the Request with the corresponding call to Core::resolve to pass the data back to the App::update function (wrapped in the event provided to the capability originating the effect).