Module crux_core::render

Built-in capability used to notify the Shell that a UI update is necessary.

Structs

Render

Use an instance of Render to notify the Shell that it should update the user interface. This assumes a declarative UI framework is used in the Shell, which will take the ViewModel provided by Core::view and reconcile the new UI state based on the view model with the previous one.

RenderOperation

The single operation Render implements.