Module crux_core::testing

Testing support for unit testing Crux apps.

Structs

AppTester

AppTester is a simplified execution environment for Crux apps for use in tests.

TestEffect

Update

Update test helper holds the result of running an app update using AppTester::update.