Trait crux_core::capability::WithContext

pub trait WithContext<App, Ef>where
    App: App,{
    fn new_with_context(
        context: CapabilityContext<Ef, App::Event>
    ) -> App::Capabilities;
}

Allows Crux to construct app’s set of required capabilities, providing context they can then use to request effects and dispatch events.

new_with_context is called by Crux and should return an instance of the app’s Capabilities type with all capabilities constructed with context passed in. Use Context::with_effect to create an appropriate context instance with the effect constructor which should wrap the requested operations.

Note that this implementation can be generated by the derive macro Effect (in the crux_macros crate).

impl crux_core::WithContext<App, Effect> for Capabilities {
    fn new_with_context(context: CapabilityContext<Effect, Event>) -> Capabilities {
        Capabilities {
            http: Http::new(context.with_effect(Effect::Http)),
            render: Render::new(context.with_effect(|_| Effect::Render)),
        }
    }
}

Required Methods

fn new_with_context(
context: CapabilityContext<Ef, App::Event>
) -> App::Capabilities