crux_core
Module compose

A capability which can spawn tasks which orchestrate across other capabilities. This is useful for orchestrating a number of different effects into a single transaction.

Structs

Compose

Compose capability can be used to orchestrate effects into a single transaction.

ComposeContext

A restricted context given to the closure passed to Compose::spawn. This context can only update the app, not request from the shell or spawn further tasks.